# documentation/openapi.yaml
openapi: 3.0.1
info:
  title: Musicman API
  description: Описание всех эндпоинтов сервиса "Musicman"
  contact:
    name: Анастасия Сашина
    email: sashina@cs.vsu.ru
  version: 1.0.6
servers:
  - url: http://localhost:8080/api/v1
    description: Generated server url
security:
  - Bearer: []
paths:
...
  "/friends/{nickname}":
    post:
      tags:
        - Friend Controller
      summary: Добавить пользователя в друзья
      operationId: addFriend
      parameters:
        - name: nickname
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Пользователь добавлен
          content:
            application/json: {}
        '400':
          description: Невалидные входные данные
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExceptionDTO"
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExceptionDTO"
        '404':
          description: Пользователя не существует
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExceptionDTO"
        '409':
          description: Пользователь уже в друзьях
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExceptionDTO"
    delete:
      tags:
        - Friend Controller
      summary: Удалить пользователя из друзей
      operationId: deleteFriend
      parameters:
        - name: nickname
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Пользователь удален из друзей
          content:
            application/json: {}
        '400':
          description: Невалидные входные данные
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExceptionDTO"
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExceptionDTO"
        '404':
          description: Пользователя не существует
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExceptionDTO"
        '405':
          description: Пользователя не было в друзьях
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExceptionDTO"
...
  "/friends":
    get:
      tags:
        - Friend Controller
      summary: Выводит всех друзей пользователя
      operationId: getAllFriends
      parameters:
        - name: pageNumber
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 1
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 5
        - name: nickname
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Возвращает всех друзей
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/FriendPagesDTO"
        '400':
          description: Невалидные входные данные
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExceptionDTO"
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExceptionDTO"
...
components:
  schemas:
    ExceptionDTO:
      type: object
      properties:
        message:
          type: string
...
    PagingInfoDTO:
      type: object
      properties:
        pageNumber:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalPages:
          type: integer
          format: int32
...
    FriendDTO:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        email:
          type: string
        nickname:
          type: string
    FriendPagesDTO:
      type: object
      properties:
        paging:
          "$ref": "#/components/schemas/PagingInfoDTO"
        data:
          type: array
          items:
            "$ref": "#/components/schemas/FriendDTO"
  securitySchemes:
    Authorization:
      type: http
      scheme: Bearer
      bearerFormat: JWT